requirements.txt
python-dotenv==1.0.1
flask==3.0.3
requests==2.32.3
prometheus-client==0.21.1
paho-mqtt==2.1.0
pytest==8.3.5
